Migrate view container component
Added aliasing for components within core, such that all imports from
src can be imported from "@/". This caused an issue in the dev
environment because files within app/ could not resolve this alias.

Instead of using @michelangelo/core in imports within packages/core,
which is pretty long, I explicitly registered the unique paths to app's
vite config. "@" was choosen because it aligns with trends in other
vite/TypeScript projects. And because packages/core will be the primary
export from our project, and the primary import into our dev
environment, app/

In the spirit of small import paths, renamed @michelangelo/gen-* to
@ma/gen-*.

With this new aliasing approach, the import sorting sourced from
internal does not correctly categorize @ma/ and @/ as "internal"
imports. Updated the regex to address this issue.

Updated a default TypeScript eslint rule that will be incompatible with
Michelangelo's open source migration -- internally we prefer Types to
Interfaces.