Replace vite alias with package.json imports
vite aliases were not sufficiently propagating across dependencies, such
that depending on an aliased-packaged required awareness of the aliases.

Refactored the approach to leverage package.json imports. package.json
imports need to be prefixed by # instead of @, so updated imports as
well and the eslint sort order configuration.

Leveraged #core instead of just #, since there will likely be more
packages in the future that would conflice the with the shared
tsconfig.base.json paths configuration.