Add query transformation to hooks (#176)
Internally, MA Studio query hooks transform Response protobuf messages
to provide direct access to the requested data. For example, instead of
getting { project: Project } as a response, the hooks provide Project
directly. Implemented this functionality for singular-entity endpoints
like Get, Create, and Update. Implemented this using
UseQueryOptions.select; removed select as an available input to
useQuery.

This updated the return type of useQuery, which means that the
QueryContextType can no longer support a generic that controls the data
result. This is because the input

Leverage prototobuf-es $typeName property that is included with
alongside the generated protobuf client code because it has all the
information required for correctly extracting the entity: endpoint name,
and entity name. In order to leverage the $typeName property, removed
the OmitTypeName wrapper from RpcResponse. $typeName is still removed
from RpcRequest. This works well because we don't require clients to
provide $typeName but benefit from the server's insertion of the
$typeName.