Refactor RPC injection solution to use functions
When integrating the RPC injection solution internally, react-compiler
lint flagged that passing hooks as props is discouraged by React.
Refactored the RPC injection solution to inject a function instead of a
hook. The function is wrapped in a hook within @michelangelo/core.

Since hooks are no longer generated by michelangelo/rpc, removed the
@tanstack/react-query dependency, in addition to the RPCProvider that
exposed QueryClientProvider and the RpcHandlers to the
@tanstack-query-wrapped hooks.

Refactored use-studio-query.test.ts to not share state between tests.
This was causing a race conditions and timeouts with tests. This is due
to tanstack/query being asynchronous whereas the previous injecting of
useQuery directly into the context was not.