Refactor RPC handlers to be a constant
The previous version for defining RPC handlers via a function felt like
a bit of an anti-pattern, since all the information is available at
compile time.

Introduced new "services.ts" to contain all the services that will be
exposed from michelangelo/rpc. This provides a clearer delineation
between handlers.ts and services.ts, and should also help reduce the
amount of relocation of "createGrpcWebTransport" and friends.

Updated project-detail to be aware of the data transformation that
happens as a result of a recent diff.