Update michelangelo-core for merging cell renderers internally (#220)
This commit includes the exporting of functionality migrated from
internal repository to this open source repository.

When testing internally relying on the built output, experienced an
issue where the build output included an attempt to import the default
export from styletron-react. It appears to be related to rollup considering
styletron-react a common JS library. Adding esmExternals: true for
styletron-react removes the attempt to import the default export from
styletron-react.

In order for the commonjsOptions to apply to styletron-react, registered
styletron-react to optimizeDeps. This appears to be by-design for vite:
https://github.com/vitejs/vite/issues/5668