Implement lightweight pipeline and project list demo (#225)
Michelangelo Studio homepage now lists projects created in the ma
sandbox. Users can click on a project to see project metadata and a list
of pipelines in that project.

Discovered that the useBinaryFormat for connect's gRPC-web client
returns enums as integers instead of strings. There is a lot of work
required to get enums as values, so leveraging text mapping exposed by
state/type cells.