Add sandbox CRDs and update sandbox creation process (#215)
Add a new 'demo' command to the sandbox tool that creates a demo project and
pipelines in the sandbox cluster. This can be used for UI sandbox users to scaffold
a sandbox that allows UI users and developers to quickly experiment with MA Studio.

The command extracts the namespace from project.yaml and creates the project
and pipelines in that namespace. In order to access the project namespace, added
pyyaml dependency for YAML parsing.