Improve icon sizing compatability (#343)
Icons provided by baseui are a factor larger than Material UI icons.
This leads to inconsistent rendering for Uber's internal application.

Apply a scale to increase the size value applied to the Material UI icon
and update instantiations of icons to user smaller sizes.

A default icon size of scale550 (default 14px) is now applied to the
Icon component, as this appears to be the common fallback for icon
sizing.