Fix table filter option generation (#345)
The initial implementation of unfiltered data resulted in columns
missing category filter options if the column used an accessor function
for getting cell values.

This is because the columns references to build the unfiltered data
list had been transformed to be Tanstack Table friendly, with accessorFn
prop instead of accessor.