Add trigger detail view with dynamic pipeline run filtering (#420)
Implement table view that displays pipeline runs related to specific
triggers.

Our gRPC client transforms protobuf oneofs to objects with "case" 
key that corresponds to the name of the oneof key.

Add interpolation support to entity detail route to resolve template
strings in service options at runtime. This was initially intended to support
dynamic serviceOptions for trigger detail page filtered, but is not needed
for this specific change. It will still be required for future use cases.